requirements.txt
python-docx>=1.1.0
docx2pdf>=0.1.8
Pillow>=10.0.0
